Completed bacckend api

+++ devforge/src/main/java/com/spring/devforge/authentication/UserDataJpa.java
@@ -2,7 +2,7 @@ package com.spring.devforge.authentication;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UserDataJpa extends JpaRepository <Users,Integer>{
+public interface UserDataJpa extends JpaRepository <Users,Long>{
 	
 	public Users findByEmail(String email);
 	public Users findByUsername(String username);
